+++ backend/src/main/java/sn/immosn/backend/client/web/annonce/dto/AnnonceListDto.java
@@ -1,10 +1,11 @@
 package sn.immosn.backend.client.web.annonce.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record AnnonceListDto (
     String libelle,
-    Double prix,
+    BigDecimal prix,
     String adresse,
     TypeBienResponseDto typeBien,
     Integer nbrPieces,
